ok, let's sdo it seriously now: use the right precision for the page duration that a PDF can give us
svn path=/trunk/playground/graphics/okular/; revision=618194

+++ core/page.cpp
@@ -78,7 +78,7 @@ class Page::Private
         PageTransition * m_transition;
         Link * m_openingAction;
         Link * m_closingAction;
-        int duration;
+        double duration;
 };
 
 void Page::Private::imageRotationDone()
@@ -420,12 +420,12 @@ void Page::setSourceReferences( const QLinkedList< SourceRefObjectRect * > refRe
         m_rects << rect;
 }
 
-void Page::setDuration( int seconds )
+void Page::setDuration( double seconds )
 {
     d->duration = seconds;
 }
 
-int Page::duration() const
+double Page::duration() const
 {
     return d->duration;
 }
